Moved movement test related files into the Movements folder
