+++ Assets/Modules/NPCs/Golden/Tests/PlayMode/Movements.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 867d0d6fd7abb0047bd044ff8cc8ca62
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
