+++ Assets/Modules/NPCs/Golden/Tests/PlayMode/Movements/MockInput.cs.meta
